+++ src/ncc/Classes/PhpExtension/PhpRuntime.php
@@ -1,33 +1,32 @@
 <?php
-/*
- * Copyright (c) Nosial 2022-2023, all rights reserved.
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- *  associated documentation files (the "Software"), to deal in the Software without restriction, including without
- *  limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- *  Software, and to permit persons to whom the Software is furnished to do so, subject to the following
- *  conditions:
- *
- *  The above copyright notice and this permission notice shall be included in all copies or substantial portions
- *  of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- *  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- *  PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- *  LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- *  OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- *  DEALINGS IN THE SOFTWARE.
- *
- */
+    /*
+     * Copyright (c) Nosial 2022-2023, all rights reserved.
+     *
+     *  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+     *  associated documentation files (the "Software"), to deal in the Software without restriction, including without
+     *  limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+     *  Software, and to permit persons to whom the Software is furnished to do so, subject to the following
+     *  conditions:
+     *
+     *  The above copyright notice and this permission notice shall be included in all copies or substantial portions
+     *  of the Software.
+     *
+     *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+     *  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+     *  PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+     *  LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+     *  OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+     *  DEALINGS IN THE SOFTWARE.
+     *
+     */
 
-namespace ncc\Classes\PhpExtension;
+    namespace ncc\Classes\PhpExtension;
 
     use Exception;
     use ncc\Enums\Options\RuntimeImportOptions;
     use ncc\Classes\NccExtension\ConstantCompiler;
-    use ncc\Exceptions\ConstantReadonlyException;
     use ncc\Exceptions\ImportException;
-    use ncc\Exceptions\InvalidConstantNameException;
+    use ncc\Exceptions\IntegrityException;
     use ncc\Interfaces\RuntimeInterface;
     use ncc\Objects\PackageLock\VersionEntry;
     use ncc\Objects\ProjectConfiguration\Assembly;
@@ -54,7 +53,9 @@ namespace ncc\Classes\PhpExtension;
             $assembly_path = $versionEntry->getInstallPaths()->getDataPath() . DIRECTORY_SEPARATOR . 'assembly';
 
             if(!file_exists($assembly_path))
+            {
                 throw new ImportException('Cannot locate assembly file \'' . $assembly_path . '\'');
+            }
 
             try
             {
@@ -85,29 +86,27 @@ namespace ncc\Classes\PhpExtension;
                     {
                         Constants::register($assembly->package, $name, $value, true);
                     }
-                    catch (ConstantReadonlyException $e)
+                    catch (IntegrityException $e)
                     {
-                        trigger_error('Constant \'' . $name . '\' is readonly (' . $assembly->package . ')', E_USER_WARNING);
-                    }
-                    catch (InvalidConstantNameException $e)
-                    {
-                        throw new ImportException('Invalid constant name \'' . $name . '\' (' . $assembly->package . ')', $e);
+                        trigger_error('Cannot set constant \'' . $name . '\', ' . $e->getMessage(), E_USER_WARNING);
                     }
                 }
             }
 
-            if(file_exists($autoload_path) && !in_array(RuntimeImportOptions::IMPORT_AUTOLOADER, $options))
+            if(file_exists($autoload_path) && !in_array(RuntimeImportOptions::IMPORT_AUTOLOADER, $options, true))
             {
                 require_once($autoload_path);
             }
 
-            if(file_exists($static_files) && !in_array(RuntimeImportOptions::IMPORT_STATIC_FILES, $options))
+            if(file_exists($static_files) && !in_array(RuntimeImportOptions::IMPORT_STATIC_FILES, $options, true))
             {
                 try
                 {
                     $static_files = ZiProto::decode(IO::fread($static_files));
                     foreach($static_files as $file)
+                    {
                         require_once($file);
+                    }
                 }
                 catch(Exception $e)
                 {
@@ -116,9 +115,6 @@ namespace ncc\Classes\PhpExtension;
 
             }
 
-            if(!file_exists($autoload_path) && !file_exists($static_files))
-                return false;
-
-            return true;
+            return !(!file_exists($autoload_path) && !file_exists($static_files));
         }
     }