+++ src/ncc/Interfaces/InstallerInterface.php
@@ -24,7 +24,6 @@ namespace ncc\Interfaces;
 
     use Exception;
     use ncc\Exceptions\ComponentChecksumException;
-    use ncc\Exceptions\ComponentDecodeException;
     use ncc\Objects\InstallationPaths;
     use ncc\Objects\Package;
     use ncc\Objects\Package\Component;
@@ -44,7 +43,6 @@ namespace ncc\Interfaces;
          * @param Component $component
          * @return string|null
          * @throws ComponentChecksumException
-         * @throws ComponentDecodeException
          */
         public function processComponent(Package\Component $component): ?string;