+++ src/ncc/Objects/Constant.php
@@ -1,28 +1,31 @@
 <?php
-/*
- * Copyright (c) Nosial 2022-2023, all rights reserved.
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- *  associated documentation files (the "Software"), to deal in the Software without restriction, including without
- *  limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
- *  Software, and to permit persons to whom the Software is furnished to do so, subject to the following
- *  conditions:
- *
- *  The above copyright notice and this permission notice shall be included in all copies or substantial portions
- *  of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
- *  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
- *  PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
- *  LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- *  OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- *  DEALINGS IN THE SOFTWARE.
- *
- */
 
-namespace ncc\Objects;
+    /** @noinspection PhpMissingFieldTypeInspection */
 
-    use ncc\Exceptions\ConstantReadonlyException;
+    /*
+     * Copyright (c) Nosial 2022-2023, all rights reserved.
+     *
+     *  Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+     *  associated documentation files (the "Software"), to deal in the Software without restriction, including without
+     *  limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+     *  Software, and to permit persons to whom the Software is furnished to do so, subject to the following
+     *  conditions:
+     *
+     *  The above copyright notice and this permission notice shall be included in all copies or substantial portions
+     *  of the Software.
+     *
+     *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+     *  INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR
+     *  PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+     *  LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+     *  OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+     *  DEALINGS IN THE SOFTWARE.
+     *
+     */
+
+    namespace ncc\Objects;
+
+    use ncc\Exceptions\IntegrityException;
     use ncc\Utilities\Resolver;
 
     class Constant
@@ -32,35 +35,35 @@ namespace ncc\Objects;
          *
          * @var string
          */
-        private $Hash;
+        private $hash;
 
         /**
          * The package that manages this constant
          *
          * @var string
          */
-        private $Scope;
+        private $scope;
 
         /**
          * The name of the constant
          *
          * @var string
          */
-        private $Name;
+        private $name;
 
         /**
          * The value of the constant
          *
          * @var string
          */
-        private $Value;
+        private $value;
 
         /**
          * Indicates if the constant is readonly or not
          *
          * @var bool
          */
-        private $Readonly;
+        private $readonly;
 
         /**
          * Public Constructor
@@ -72,11 +75,11 @@ namespace ncc\Objects;
          */
         public function __construct(string $scope, string $name, string $value, bool $readonly=false)
         {
-            $this->Scope = $scope;
-            $this->Name = $name;
-            $this->Value = $value;
-            $this->Readonly = $readonly;
-            $this->Hash = Resolver::resolveConstantHash($this->Scope, $this->Name);
+            $this->scope = $scope;
+            $this->name = $name;
+            $this->value = $value;
+            $this->readonly = $readonly;
+            $this->hash = Resolver::resolveConstantHash($this->scope, $this->name);
         }
 
         /**
@@ -86,7 +89,7 @@ namespace ncc\Objects;
          */
         public function __toString(): string
         {
-            return $this->Value;
+            return $this->value;
         }
 
         /**
@@ -94,7 +97,7 @@ namespace ncc\Objects;
          */
         public function getValue(): string
         {
-            return $this->Value;
+            return $this->value;
         }
 
         /**
@@ -104,23 +107,23 @@ namespace ncc\Objects;
          */
         public function getFullName(): string
         {
-            return Resolver::resolveFullConstantName($this->Scope, $this->Name);
+            return Resolver::resolveFullConstantName($this->scope, $this->name);
         }
 
         /**
          * @param string $value
          * @param bool $readonly
-         * @throws ConstantReadonlyException
+         * @throws IntegrityException
          */
         public function setValue(string $value, bool $readonly=false): void
         {
-            if($this->Readonly)
+            if($this->readonly)
             {
-                throw new ConstantReadonlyException('Cannot set value to the constant \'' .  $this->getFullName() .  '\', constant is readonly');
+                throw new IntegrityException('Cannot set value to the constant \'' .  $this->getFullName() .  '\', constant is readonly');
             }
 
-            $this->Value = $value;
-            $this->Readonly = $readonly;
+            $this->value = $value;
+            $this->readonly = $readonly;
         }
 
         /**
@@ -128,7 +131,7 @@ namespace ncc\Objects;
          */
         public function isReadonly(): bool
         {
-            return $this->Readonly;
+            return $this->readonly;
         }
 
         /**
@@ -136,7 +139,7 @@ namespace ncc\Objects;
          */
         public function getHash(): string
         {
-            return $this->Hash;
+            return $this->hash;
         }
 
         /**
@@ -144,7 +147,7 @@ namespace ncc\Objects;
          */
         public function getScope(): string
         {
-            return $this->Scope;
+            return $this->scope;
         }
 
         /**
@@ -152,6 +155,6 @@ namespace ncc\Objects;
          */
         public function getName(): string
         {
-            return $this->Name;
+            return $this->name;
         }
     }